feat: update repository deal with error

+++ backend/capstone/src/main/java/com/capstone/repository/EquipmentRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.capstone.model.Equipment;
 
 @Repository
-public interface EquipmentRepository extends JpaRepository<Equipment, Long> {
+public interface EquipmentRepository extends JpaRepository<Equipment, Integer> {
 
 }
