+++ backend/capstone/src/main/java/com/capstone/repository/RecordRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.capstone.model.Record;
 
 @Repository
-public interface RecordRepository extends JpaRepository<Record, Long>{
+public interface RecordRepository extends JpaRepository<Record, Integer>{
 
 }
